[network] Fixed No IPv6 issue and rewrote check
In the first place, I just fixed a bug that was making gohai fail
completely when no IPv6 address was set on the Machine. I actually ended
up rewriting the whole network check for Windows because a couple of
things didn't make sense at all (for instance, we could end up sending
the Mac Address of a given interface, the IPv4 of another one and the
IPv6 of another again... I guess we absolutely need to avoid that in the
future).

I also added a test for my change. Run it with

```
go test *_windows.go *_windows_test.go
```

It was a real pain since I'm using the output of ipconfig on Windows for
that test but \x00 are appended after every character in my text (I am
sure it's an encoding issue but I didn't find a proper way to solve it
once and for all... apart from this hack).